implementation/requirements.txt
cs50
csv
datetime
flask
flask_session
os
requests
SPARQLWrapper
tempfile
threading
werkzeug.utils